first major code refactoring
* one lib for nconf defaults
* one lib for configuration options
* one lib for message management
* one lib for schema management
* one lib for checks management
